requirements.txt
fastapi==0.73.0
fuzzywuzzy==0.18.0
gspread==4.0.1
passlib==1.7.4
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
SQLAlchemy==1.4.31
SQLAlchemy-serializer==1.4.1
uvicorn==0.17.4
gunicorn==20.1.0
python-dotenv==0.19.2
bscrypt==3.2.0